requirements.txt
streamlit==1.27.0
pandas==2.0.3
openai==1.3.0
python-dotenv==1.0.0
openpyxl==3.1.2
xlrd==2.0.1
requests==2.31.0
numpy==1.24.3
jsonschema==4.17.3
